requirements.txt
-i https://pypi.org/simple
certifi==2018.11.29
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
idna==2.5
msgpack-numpy==0.4.3.2
msgpack==0.5.6
murmurhash==1.0.1
newsapi-python==0.2.3
numpy==1.15.4
plac==0.9.6
preshed==2.0.1
pyreadline==2.1
regex==2018.1.10
requests==2.17.1
six==1.11.0
spacy==2.0.18
thinc==6.12.1
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
urllib3==1.21.1
wrapt==1.10.11
